gdb/ 	* corefile.c (read_memory, read_stack, write_memory): Accept LEN 	argument as ssize_t. 	* gdbcore.h (read_memory, read_stack, write_memory): Likewise. 	* remote.c (remote_write_bytes_aux, remote_write_bytes): Likewise. 	* target.c (target_read_stack, target_write_memory) 	(target_write_raw_memory): Likewise. 	* target.h (target_read_stack, target_write_memory) 	(target_write_raw_memory): Likewise.

+++ gdb/corefile.c
@@ -213,7 +213,7 @@ memory_error (int status, CORE_ADDR memaddr)
 /* Same as target_read_memory, but report an error if can't read.  */
 
 void
-read_memory (CORE_ADDR memaddr, gdb_byte *myaddr, int len)
+read_memory (CORE_ADDR memaddr, gdb_byte *myaddr, ssize_t len)
 {
   int status;
 
@@ -225,7 +225,7 @@ read_memory (CORE_ADDR memaddr, gdb_byte *myaddr, int len)
 /* Same as target_read_stack, but report an error if can't read.  */
 
 void
-read_stack (CORE_ADDR memaddr, gdb_byte *myaddr, int len)
+read_stack (CORE_ADDR memaddr, gdb_byte *myaddr, ssize_t len)
 {
   int status;
 
@@ -352,7 +352,7 @@ read_memory_typed_address (CORE_ADDR addr, struct type *type)
    write.  */
 void
 write_memory (CORE_ADDR memaddr, 
-	      const bfd_byte *myaddr, int len)
+	      const bfd_byte *myaddr, ssize_t len)
 {
   int status;
